Consolidate dashboard modules
