--- src/introspection_dashboard.py
@@ -1,125 +1,0 @@
-from __future__ import annotations
-
-import json
-from http.server import BaseHTTPRequestHandler
-from typing import Any, Dict, Type
-
-try:
-    from .dashboard_import_helper import load_base_dashboard
-except Exception:  # pragma: no cover - fallback when not packaged
-    import importlib.util
-    import sys
-    from pathlib import Path
-
-    spec = importlib.util.spec_from_file_location(
-        "dashboard_import_helper", Path(__file__).with_name("dashboard_import_helper.py")
-    )
-    module = importlib.util.module_from_spec(spec)
-    assert spec.loader is not None
-    spec.loader.exec_module(module)  # type: ignore
-    sys.modules.setdefault("dashboard_import_helper", module)
-    load_base_dashboard = module.load_base_dashboard  # type: ignore
-
-BaseDashboard = load_base_dashboard(__file__)
-
-from .graph_of_thought import GraphOfThought
-from .reasoning_history import ReasoningHistoryLogger
-from .telemetry import TelemetryLogger
-
-
-class IntrospectionDashboard(BaseDashboard):
-    """Serve reasoning history merged with telemetry statistics."""
-
-    def __init__(
-        self,
-        graph: GraphOfThought,
-        history: ReasoningHistoryLogger,
-        telemetry: TelemetryLogger,
-    ) -> None:
-        super().__init__()
-        self.graph = graph
-        self.history = history
-        self.telemetry = telemetry
-
-    # --------------------------------------------------------------
-    def aggregate(self) -> Dict[str, Any]:
-        """Return graph JSON, history log and telemetry metrics."""
-        return {
-            "graph": self.graph.to_json(),
-            "history": self.history.get_history(),
-            "telemetry": self.telemetry.get_stats(),
-        }
-
-    # --------------------------------------------------------------
-    def to_html(self) -> str:
-        data = self.aggregate()
-        tele = data.get("telemetry", {})
-        rows = "".join(
-            f"<tr><td>{k}</td><td>{float(v):.2f}</td></tr>"
-            for k, v in tele.items()
-            if isinstance(v, (int, float))
-        )
-        hist_rows = "".join(
-            f"<li>{ts}: {summary}</li>" for ts, summary in data.get("history", [])[-5:]
-        )
-        return (
-            "<html><body><h1>Introspection Dashboard</h1>"
-            "<h2>Telemetry</h2><table border='1'>"
-            "<tr><th>Metric</th><th>Value</th></tr>"
-            f"{rows}</table>"
-            "<h2>History</h2><ul>" + hist_rows + "</ul>"
-            "</body></html>"
-        )
-
-    # --------------------------------------------------------------
-    def get_handler(self) -> Type[BaseHTTPRequestHandler]:
-        dashboard = self
-
-        class Handler(BaseHTTPRequestHandler):
-            def do_GET(self) -> None:  # noqa: D401
-                if self.path == "/graph":
-                    data = json.dumps(dashboard.graph.to_json()).encode()
-                    self.send_response(200)
-                    self.send_header("Content-Type", "application/json")
-                    self.end_headers()
-                    self.wfile.write(data)
-                elif self.path == "/history":
-                    data = json.dumps(dashboard.history.get_history()).encode()
-                    self.send_response(200)
-                    self.send_header("Content-Type", "application/json")
-                    self.end_headers()
-                    self.wfile.write(data)
-                elif self.path in ("/stats", "/telemetry"):
-                    data = json.dumps(dashboard.telemetry.get_stats()).encode()
-                    self.send_response(200)
-                    self.send_header("Content-Type", "application/json")
-                    self.end_headers()
-                    self.wfile.write(data)
-                elif self.path == "/json":
-                    data = json.dumps(dashboard.aggregate()).encode()
-                    self.send_response(200)
-                    self.send_header("Content-Type", "application/json")
-                    self.end_headers()
-                    self.wfile.write(data)
-                else:
-                    html = dashboard.to_html().encode()
-                    self.send_response(200)
-                    self.send_header("Content-Type", "text/html")
-                    self.end_headers()
-                    self.wfile.write(html)
-
-            def log_message(self, format: str, *args: Any) -> None:  # noqa: D401
-                return
-
-        return Handler
-
-    def start(self, host: str = "localhost", port: int = 8060) -> None:
-        super().start(host, port)
-
-    # --------------------------------------------------------------
-    def stop(self) -> None:
-        super().stop()
-
-
-__all__ = ["IntrospectionDashboard"]
-
